requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
bcrypt
openai
pymupdf
openpyxl
python-multipart
boto3
pdfplumber
requests
pydantic

Pillow
reportlab
